create hash submodule for block
